pyproject.toml
[project]
name = "bot-ferm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.3",
    "greenlet>=3.2.4",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.12.4",
    "pytest>=9.0.1",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
